fix(common): use size_t instead of int

+++ includes/common/parser.hpp
@@ -19,7 +19,7 @@ inline Command parse_command(const std::string& line) {
     auto words = split(line);
     result.timestamp = to_int(words[0].substr(1, words[0].size() - 2));
     result.name = words[1];
-    for (int i = 2; i < words.size(); i += 2) {
+    for (size_t i = 2; i < words.size(); i += 2) {
         result.args[words[i][1] - 'a'] = words[i + 1];
     }
     return result;
